Proper fix for `keep_failed_instances`

+++ hsdp/resource_container_host.go
@@ -402,7 +402,10 @@ func resourceContainerHostCreate(ctx context.Context, d *schema.ResourceData, m 
 	_, err = stateConf.WaitForStateContext(ctx)
 	if err != nil {
 		// Trigger a delete to prevent failed instances from lingering
-		_, _, _ = client.Destroy(tagName)
+		if !keepFailedInstances {
+			_, _, _ = client.Destroy(tagName)
+			d.SetId("")
+		}
 		return diag.FromErr(fmt.Errorf(
 			"error waiting for instance '%s' to become ready: %s",
 			instanceID, err))
@@ -789,7 +792,6 @@ func resourceContainerHostDelete(_ context.Context, d *schema.ResourceData, m in
 	}
 
 	tagName := d.Get("name").(string)
-	keepFailedInstances := d.Get("keep_failed_instances").(bool)
 
 	ch, _, err := client.GetDetails(tagName)
 	if err != nil {
@@ -797,11 +799,6 @@ func resourceContainerHostDelete(_ context.Context, d *schema.ResourceData, m in
 	}
 	if ch.InstanceID != d.Id() {
 		return diag.FromErr(ErrInstanceIDMismatch)
-	}
-	state, _, err := client.GetDeploymentState(tagName)
-	if err == nil && state == "failed" && keepFailedInstances { // Keep
-		d.SetId("")
-		return diags
 	}
 	_, _, err = client.Destroy(tagName)
 	if err != nil {
